requirements.txt
altgraph==0.17.2
appscript==1.2.0
async-generator==1.10
attrs==21.4.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
colour==0.1.5
cryptography==36.0.1
future==0.18.2
h11==0.13.0
idna==3.3
imutils==0.5.4
Jinja2==3.1.1
joblib==1.1.0
junos-eznc==2.6.3
lxml==4.8.0
macholib==1.16
MarkupSafe==2.1.1
ncclient==0.6.9
netaddr==0.8.0
netmiko==4.0.0
ntc-templates==3.0.0
numpy==1.22.2
opencv-python==4.5.5.62
outcome==1.1.0
paramiko==2.10.4
Pillow==9.0.1
PyAudio==0.2.11
pycparser==2.21
pyinstaller==5.1
pyinstaller-hooks-contrib==2022.7
PyNaCl==1.5.0
pyobjc==8.3
pyobjc-core==8.3
pyobjc-framework-Accessibility==8.3
pyobjc-framework-Accounts==8.3
pyobjc-framework-AddressBook==8.3
pyobjc-framework-AdServices==8.3
pyobjc-framework-AdSupport==8.3
pyobjc-framework-AppleScriptKit==8.3
pyobjc-framework-AppleScriptObjC==8.3
pyobjc-framework-ApplicationServices==8.3
pyobjc-framework-AppTrackingTransparency==8.3
pyobjc-framework-AudioVideoBridging==8.3
pyobjc-framework-AuthenticationServices==8.3
pyobjc-framework-AutomaticAssessmentConfiguration==8.3
pyobjc-framework-Automator==8.3
pyobjc-framework-AVFoundation==8.3
pyobjc-framework-AVKit==8.3
pyobjc-framework-BusinessChat==8.3
pyobjc-framework-CalendarStore==8.3
pyobjc-framework-CallKit==8.3
pyobjc-framework-CFNetwork==8.3
pyobjc-framework-ClassKit==8.3
pyobjc-framework-CloudKit==8.3
pyobjc-framework-Cocoa==8.3
pyobjc-framework-Collaboration==8.3
pyobjc-framework-ColorSync==8.3
pyobjc-framework-Contacts==8.3
pyobjc-framework-ContactsUI==8.3
pyobjc-framework-CoreAudio==8.3
pyobjc-framework-CoreAudioKit==8.3
pyobjc-framework-CoreBluetooth==8.3
pyobjc-framework-CoreData==8.3
pyobjc-framework-CoreHaptics==8.3
pyobjc-framework-CoreLocation==8.3
pyobjc-framework-CoreMedia==8.3
pyobjc-framework-CoreMediaIO==8.3
pyobjc-framework-CoreMIDI==8.3
pyobjc-framework-CoreML==8.3
pyobjc-framework-CoreMotion==8.3
pyobjc-framework-CoreServices==8.3
pyobjc-framework-CoreSpotlight==8.3
pyobjc-framework-CoreText==8.3
pyobjc-framework-CoreWLAN==8.3
pyobjc-framework-CryptoTokenKit==8.3
pyobjc-framework-DeviceCheck==8.3
pyobjc-framework-DictionaryServices==8.3
pyobjc-framework-DiscRecording==8.3
pyobjc-framework-DiscRecordingUI==8.3
pyobjc-framework-DiskArbitration==8.3
pyobjc-framework-DVDPlayback==8.3
pyobjc-framework-EventKit==8.3
pyobjc-framework-ExceptionHandling==8.3
pyobjc-framework-ExecutionPolicy==8.3
pyobjc-framework-ExternalAccessory==8.3
pyobjc-framework-FileProvider==8.3
pyobjc-framework-FileProviderUI==8.3
pyobjc-framework-FinderSync==8.3
pyobjc-framework-FSEvents==8.3
pyobjc-framework-GameCenter==8.3
pyobjc-framework-GameController==8.3
pyobjc-framework-GameKit==8.3
pyobjc-framework-GameplayKit==8.3
pyobjc-framework-ImageCaptureCore==8.3
pyobjc-framework-IMServicePlugIn==8.3
pyobjc-framework-InputMethodKit==8.3
pyobjc-framework-InstallerPlugins==8.3
pyobjc-framework-InstantMessage==8.3
pyobjc-framework-Intents==8.3
pyobjc-framework-IOSurface==8.3
pyobjc-framework-iTunesLibrary==8.3
pyobjc-framework-KernelManagement==8.3
pyobjc-framework-LatentSemanticMapping==8.3
pyobjc-framework-LaunchServices==8.3
pyobjc-framework-libdispatch==8.3
pyobjc-framework-LinkPresentation==8.3
pyobjc-framework-LocalAuthentication==8.3
pyobjc-framework-MapKit==8.3
pyobjc-framework-MediaAccessibility==8.3
pyobjc-framework-MediaLibrary==8.3
pyobjc-framework-MediaPlayer==8.3
pyobjc-framework-MediaToolbox==8.3
pyobjc-framework-Metal==8.3
pyobjc-framework-MetalKit==8.3
pyobjc-framework-MetalPerformanceShaders==8.3
pyobjc-framework-MetalPerformanceShadersGraph==8.3
pyobjc-framework-MLCompute==8.3
pyobjc-framework-ModelIO==8.3
pyobjc-framework-MultipeerConnectivity==8.3
pyobjc-framework-NaturalLanguage==8.3
pyobjc-framework-NetFS==8.3
pyobjc-framework-Network==8.3
pyobjc-framework-NetworkExtension==8.3
pyobjc-framework-NotificationCenter==8.3
pyobjc-framework-OpenDirectory==8.3
pyobjc-framework-OSAKit==8.3
pyobjc-framework-OSLog==8.3
pyobjc-framework-PassKit==8.3
pyobjc-framework-PencilKit==8.3
pyobjc-framework-Photos==8.3
pyobjc-framework-PhotosUI==8.3
pyobjc-framework-PreferencePanes==8.3
pyobjc-framework-PushKit==8.3
pyobjc-framework-Quartz==8.3
pyobjc-framework-QuickLookThumbnailing==8.3
pyobjc-framework-ReplayKit==8.3
pyobjc-framework-SafariServices==8.3
pyobjc-framework-SceneKit==8.3
pyobjc-framework-ScreenSaver==8.3
pyobjc-framework-ScreenTime==8.3
pyobjc-framework-ScriptingBridge==8.3
pyobjc-framework-SearchKit==8.3
pyobjc-framework-Security==8.3
pyobjc-framework-SecurityFoundation==8.3
pyobjc-framework-SecurityInterface==8.3
pyobjc-framework-ServiceManagement==8.3
pyobjc-framework-Social==8.3
pyobjc-framework-SoundAnalysis==8.3
pyobjc-framework-Speech==8.3
pyobjc-framework-SpriteKit==8.3
pyobjc-framework-StoreKit==8.3
pyobjc-framework-SyncServices==8.3
pyobjc-framework-SystemConfiguration==8.3
pyobjc-framework-SystemExtensions==8.3
pyobjc-framework-UniformTypeIdentifiers==8.3
pyobjc-framework-UserNotifications==8.3
pyobjc-framework-UserNotificationsUI==8.3
pyobjc-framework-VideoSubscriberAccount==8.3
pyobjc-framework-VideoToolbox==8.3
pyobjc-framework-Virtualization==8.3
pyobjc-framework-Vision==8.3
pyobjc-framework-WebKit==8.3
pyOpenSSL==22.0.0
pyparsing==3.0.8
PyQt5==5.15.4
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
pyqt5-tools==5.15.4.3.2
pyserial==3.5
PySocks==1.7.1
python-dotenv==0.20.0
pyttsx3==2.90
PyYAML==6.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
regex==2022.3.2
requests==2.27.1
scp==0.14.4
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SpeechRecognition==3.8.1
tenacity==8.0.1
textfsm==1.1.2
tk==0.1.0
tkmacosx==1.0.4
tqdm==4.63.0
transitions==0.8.11
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.1.0
yamlordereddictloader==0.4.0
